+++ src/Models/Filepond.php
@@ -16,7 +16,7 @@ use Illuminate\Database\Eloquent\SoftDeletes;
  * @property string $filepath
  * @property string $extension
  * @property string $mimetype
- * @property string $metadata
+ * @property array $metadata
  * @property string $disk
  * @property string $upload_id
  * @property array $upload_tags
@@ -33,6 +33,7 @@ class Filepond extends Model
     protected $guarded = [];
 
     protected $casts = [
+        'metadata' => 'json',
         'upload_tags' => 'json',
         'expires_at' => 'datetime',
     ];