+++ tests/Feature/FilepondFacadeTest.php
@@ -37,6 +37,71 @@ class FilepondFacadeTest extends TestCase
     }
 
     #[Test]
+    public function can_get_metadata_after_file_upload()
+    {
+        Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));
+
+        $user = User::factory()->create();
+
+        $uploadedFile = UploadedFile::fake()->image('avatar.png', 1024, 1024);
+
+        $metadata = ['some' => 'value'];
+
+        $response = $this
+            ->actingAs($user)
+            ->call(
+                method: 'POST',
+                uri: route('filepond-process'),
+                parameters: ['avatar' => json_encode($metadata)],
+                files: ['avatar' => $uploadedFile],
+                server: $this->transformHeadersToServerVars([
+                    'Content-Type' => 'multipart/form-data',
+                    'Accept' => 'application/json',
+                ])
+            );
+
+        $responseMetadata = Filepond::field($response->content())->getMetadata();
+
+        $this->assertEquals($metadata, $responseMetadata);
+    }
+
+    #[Test]
+    public function can_get_metadata_after_multiple_file_upload()
+    {
+        Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));
+
+        $user = User::factory()->create();
+
+        $responses = [];
+        $metadatas = [];
+
+        // Create 5 temporary file uploads
+        for ($i = 1; $i <= 5; $i++) {
+            $metadata = ['some' => 'value-'.$i];
+
+            $response = $this
+                ->actingAs($user)
+                ->call(
+                    method: 'POST',
+                    uri: route('filepond-process'),
+                    parameters: ['gallery' => json_encode($metadata)],
+                    files: ['gallery' => UploadedFile::fake()->image('gallery-'.$i.'.png', 1024, 1024)],
+                    server: $this->transformHeadersToServerVars([
+                        'Content-Type' => 'multipart/form-data',
+                        'Accept' => 'application/json',
+                    ])
+                );
+
+            $responses[] = $response->content();
+            $metadatas[] = $metadata;
+        }
+
+        $responseMetadata = Filepond::field($responses)->getMetadata();
+
+        $this->assertEquals($metadatas, $responseMetadata);
+    }
+
+    #[Test]
     public function can_copy_file_upload_to_desired_location()
     {
         Storage::disk(config('filepond.temp_disk', 'local'))->deleteDirectory(config('filepond.temp_folder', 'filepond/temp'));