+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `laravel-filepond` will be documented in this file.
 
+## 13.1.1 - 2026-03-30
+
+- Support for [filepond-plugin-file-metadata](https://pqina.nl/filepond/docs/api/plugins/file-metadata/) added. ✨
+- New method `Filepond::field()->getMetadata()` added. ✨
+- New test cases added to support above changes. 🧪
+
 ## 13.0.1 - 2026-03-28
 
 - Mimetype missing after chunk upload fixed. 🐛